Move `map` folder into `game`
