+++ packages/client/src/network/connection.ts
@@ -5,7 +5,6 @@ import { Modules, Opcodes, Packets } from '@kaetram/common/network';
 
 import log from '../lib/log';
 import * as Detect from '../utils/detect';
-import TeamWar from './impl/teamwar';
 
 import type {
     CombatHitData,
@@ -63,8 +62,6 @@ export default class Connection {
     private info;
     private pointer;
     // private inventory = this.game.inventory;
-
-    private teamWar = new TeamWar();
 
     private population!: number;
     private queueColour!: string | null;
