+++ packages/server/src/game/map/areas/impl/chest.ts
@@ -1,6 +1,6 @@
 import Utils from '@kaetram/common/util/utils';
 
-import World from '../../../game/world';
+import World from '../../../world';
 import Area from '../area';
 import Areas from '../areas';
