+++ packages/server/src/game/map/areas/index.ts
