+++ packages/server/src/game/world.ts
@@ -7,8 +7,8 @@ import log from '@kaetram/common/util/log';
 import Entities from '../controllers/entities';
 import GlobalObjects from '../controllers/globalobjects';
 import Shops from '../controllers/shops';
-import Grids from '../map/grids';
-import Map from '../map/map';
+import Grids from './map/grids';
+import Map from './map/map';
 import API from '../network/api';
 import Discord from '../network/discord';
 import Messages, { Packet } from '../network/messages';
